emma cropped to fit a circle (not ellipse)
I don't like it, but it works, and seems to be the only solution that
fits all requirements without using js:

For the border(-radius) to form a circle, not ellipse, the element has
to be square. It should also not be bigger than either an arbitrary
width or height, whichever is smallest, all while maintaining its aspect
ratio.