Replace Pipenv with pip-tools
There has been no release of Pipenv for over a year and little
communication was made with the community during this time. Now a
maintainer has said that a release is soon to be tagged, but I feel like
they've shown that they are unable to run such a popular project,
despite many people making PRs and offering to help.

Also, pip-tools hopefully won't get murdered by the OOM-killer on low
resource servers, unlike Pipenv.