Normalize phone numbers in the database.
Add form validation to ensure valid phone numbers are entered, and a
custom setter in the User model, automatically normalizing numbers when
they are set.

Further, a script is added to normalize phone numbers in an existing
database. Simply run `flask scripts normalize_phone_numbers` to execute
it.

Additionally, update the test database to have valid but still unused
phone numbers, as well as one user without a phone number. The old
numbers for Monty Python and Rick Astley used the 071 area code, which
the phonenumber library doesn't like. The replacement numbers are taken
from the [PTS list of phone numbers to use in books, movies,
etc](https://www.pts.se/sv/bransch/telefoni/nummer-och-adressering/telefonnummer-for-anvandning-i-bocker-och-filmer-etc/).
:)

The purpose of this change is to enable searching for users by phone
number, which will be needed for integration with
[Sqrubbentelefonen](https://github.com/Sqrubben/Sqrubbentelefonen).