Fix bug in image url generation (discarded width).
The image urls were previously generated using the werkzeug.urls module.
However, those methods were deprecated and eventually removed in
werkzeug 2.3, see https://github.com/pallets/werkzeug/issues/2600.

When migrating to instead use urllib.parse, there was in error in the
usage of urljoin (a missing trailing slash) that caused the /imgXXX/
part of the url to be discarded.